Fix warning.

+++ src/sbbs3/js_bbs.cpp
@@ -1390,7 +1390,7 @@ js_atcode(JSContext *cx, uintN argc, jsval *arglist)
 	sbbs_t*		sbbs;
 	char	str[128],str2[128],*p;
 	char	*instr;
-	int		disp_len;
+	size_t	disp_len;
 	bool	padded_left=false;
 	bool	padded_right=false;
 	const char *cp;
